requirements.txt
# Backend dependencies
fastapi>=0.104.0
uvicorn>=0.23.2
python-dotenv>=1.0.0
pydantic>=2.4.2
google-generativeai>=0.3.1
huggingface-hub>=0.19.4
requests>=2.31.0
python-multipart>=0.0.6

# Frontend dependencies
streamlit>=1.28.0
streamlit-chat>=0.1.1
matplotlib>=3.8.0
seaborn>=0.13.0
reportlab>=4.0.7

# Voice capabilities
SpeechRecognition>=3.10.0
gTTS>=2.4.0
pyaudio>=0.2.13

# Data handling
pandas>=2.1.1
numpy>=1.26.0

fastapi==0.104.1
uvicorn==0.23.2
requests==2.31.0
streamlit==1.28.1
langchain==0.0.335
python-dotenv==1.0.0
SpeechRecognition==3.14.3
gtts==2.5.4
reportlab==4.0.7
pydantic==2.4.2
huggingface-hub==0.19.4
transformers==4.35.2
pyaudio==0.2.13
numpy==1.26.1
pandas==2.1.3
matplotlib==3.8.1
seaborn==0.13.2
google-generativeai==0.3.1
torch==2.7.1
torchvision==0.22.1
torchaudio==2.7.1
accelerate==1.9.0
bitsandbytes==0.41.1